patch 9.1.0545: MSVC conversion warning
Problem:  MSVC conversion warning (LemonBoy, after 9.1.0522)
Solution: Use size_t instead of int, fix style of casts
          (Ernie Rael)

related: #15082
closes: #15187

+++ src/eval.c
@@ -6197,10 +6197,10 @@ object_tv2string(
     else if (copyID != 0 && obj->obj_copyID == copyID
             && obj->obj_class->class_obj_member_count != 0)
     {
-	int n = 25 + strlen((char*)obj->obj_class->class_name);
+	size_t n = 25 + strlen((char *)obj->obj_class->class_name);
 	r = alloc(n);
 	if (r != NULL)
-	    (void)vim_snprintf((char*)r, n, "object of %s {...}",
+	    (void)vim_snprintf((char *)r, n, "object of %s {...}",
 						obj->obj_class->class_name);
 	*tofree = r;
     }
